requirements.txt
# requirements.txt
openai-whisper==20231117
torch>=2.0.0
numpy>=1.20.0
pydub>=0.25.1
python-dotenv>=1.0.0